Removed double and replaced with float

+++ league_results_processor/dtos/game_points.py
@@ -1,15 +1,12 @@
-from tokenize import Double
-
-
 class GamePoints(object):
     """docstring for GamePoints."""
-    def __init__(self, win_points: Double, draw_points: Double, loss_points: Double):
+    def __init__(self, win_points: float, draw_points: float, loss_points: float):
         self.win_points = win_points
         self.draw_points = draw_points
         self.loss_points = loss_points
 
     @property
-    def win_points(self) -> Double:
+    def win_points(self) -> float:
         return self._win_points
 
     @win_points.setter
@@ -17,7 +14,7 @@ class GamePoints(object):
         self._win_points = value
 
     @property
-    def draw_points(self) -> Double:
+    def draw_points(self) -> float:
         return self._draw_points
 
     @draw_points.setter
@@ -25,7 +22,7 @@ class GamePoints(object):
         self._draw_points = value
 
     @property
-    def loss_points(self) -> Double:
+    def loss_points(self) -> float:
         return self._loss_points
 
     @loss_points.setter
